Update- hacer los domulos segun la guia del ejercio
