+++ Semana 10/main.py
@@ -1,0 +1,5 @@
+from menu import menu
+
+
+while menu():
+    pass
