--- Semana 10/main_week10.py
@@ -1,5 +1,0 @@
-from menu_options.main_menu import menu
-
-
-while menu():
-    pass
