--- Semana 10/menu_options/student_list.csv
@@ -1,11 +1,0 @@
-name,last_name,section,grade_spanish,grade_english,grade_socials,grade_science
-Alice,Wonderland,A,90,85,92,88
-Bob,Marley,B,78,82,75,80
-Charlie,Brown,A,95,90,98,92
-David,Copperfield,B,85,88,82,89
-Emily,Dickinson,A,88,92,85,90
-Frank,Sinatra,B,72,75,70,78
-Grace,Kelly,A,91,89,90,87
-Harry,Potter,B,80,85,78,82
-Isabella,Swan,A,86,91,88,90
-Jack,Sparrow,B,75,80,72,78
