--- Semana 10/menu_options/test
@@ -1,92 +1,0 @@
-students = [
-    {
-        'name': 'Alice',
-        'last_name': 'Wonderland',
-        'section': 'A',
-        'grade_spanish': 90,
-        'grade_english': 85,
-        'grade_socials': 92,
-        'grade_science': 88,
-    },
-    {
-        'name': 'Bob',
-        'last_name': 'Marley',
-        'section': 'B',
-        'grade_spanish': 78,
-        'grade_english': 82,
-        'grade_socials': 75,
-        'grade_science': 80,
-    },
-    {
-        'name': 'Charlie',
-        'last_name': 'Brown',
-        'section': 'A',
-        'grade_spanish': 95,
-        'grade_english': 90,
-        'grade_socials': 98,
-        'grade_science': 92,
-    },
-    {
-        'name': 'David',
-        'last_name': 'Copperfield',
-        'section': 'B',
-        'grade_spanish': 85,
-        'grade_english': 88,
-        'grade_socials': 82,
-        'grade_science': 89,
-    },
-    {
-        'name': 'Emily',
-        'last_name': 'Dickinson',
-        'section': 'A',
-        'grade_spanish': 88,
-        'grade_english': 92,
-        'grade_socials': 85,
-        'grade_science': 90,
-    },
-    {
-        'name': 'Frank',
-        'last_name': 'Sinatra',
-        'section': 'B',
-        'grade_spanish': 72,
-        'grade_english': 75,
-        'grade_socials': 70,
-        'grade_science': 78,
-    },
-    {
-        'name': 'Grace',
-        'last_name': 'Kelly',
-        'section': 'A',
-        'grade_spanish': 91,
-        'grade_english': 89,
-        'grade_socials': 90,
-        'grade_science': 87,
-    },
-    {
-        'name': 'Harry',
-        'last_name': 'Potter',
-        'section': 'B',
-        'grade_spanish': 80,
-        'grade_english': 85,
-        'grade_socials': 78,
-        'grade_science': 82,
-    },
-    {
-        'name': 'Isabella',
-        'last_name': 'Swan',
-        'section': 'A',
-        'grade_spanish': 86,
-        'grade_english': 91,
-        'grade_socials': 88,
-        'grade_science': 90,
-    },
-    {
-        'name': 'Jack',
-        'last_name': 'Sparrow',
-        'section': 'B',
-        'grade_spanish': 75,
-        'grade_english': 80,
-        'grade_socials': 72,
-        'grade_science': 78,
-    },
-]
